requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
asyncpg==0.30.0
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.8
cohere==5.13.4
coloredlogs==15.0.1
dataclasses-json==0.6.7
distro==1.9.0
fastapi==0.115.6
fastavro==1.10.0
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.12.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.0
humanfriendly==10.0
idna==3.10
jiter==0.8.2
jq==1.8.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.13
langchain-cohere==0.3.4
langchain-community==0.3.13
langchain-core==0.3.28
langchain-experimental==0.3.4
langchain-openai==0.2.14
langchain-postgres==0.0.12
langchain-redis==0.1.1
langchain-text-splitters==0.3.4
langsmith==0.2.4
marshmallow==3.23.2
ml-dtypes==0.4.1
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.58.1
orjson==3.10.12
packaging==24.2
pandas==2.2.3
parameterized==0.9.0
pgvector==0.2.5
propcache==0.2.1
psycopg==3.2.3
psycopg-pool==3.2.4
psycopg2-binary==2.9.10
pydantic==2.10.4
pydantic-settings==2.7.0
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-ulid==2.7.0
pytz==2024.2
PyYAML==6.0.2
rank-bm25==0.2.2
redis==5.2.1
redisvl==0.3.7
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
tabulate==0.9.0
tenacity==9.0.0
tiktoken==0.8.0
tokenizers==0.21.0
tqdm==4.67.1
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
yarl==1.18.3
